evt: add event server and satellite that manage client subscriptions
The satellite can have authoritative models that can be published locally
and immediately and written to the database. If the server is online the
transactions relayed and on response replicated and the local cleared
If the server was offline the client sends all local transactions to the
server on connect and again receives the responses and clears locals.