wshub: increase msg write timeout again, use different timeouts cli and srv
My migration replay script for satellites pulls a lot of events.
This works fine on my workstation with m2 discs but on the raspberry pi.

My client usually should not send big message and should not wait to long
as not to interfere with the reconnect logic. The server might however
send large message and it is no problem if one client hangs.