wshub: add hack to send websocket messages as binary
the whole websocket binary/text thing, is basically moot. however, in the
browser binary messages are received as array buffer while text messages
are received as utf-8 bytes converted to ucs-2 js strings. this is
basically the only difference. if i were to choose all messages should be
binary, because it is easier to convert a array buffer to a string in
javascript than the other way round.