dom: introduce dom mod and use custom dom spec base
As it turns out the requirements for the dom specs are not met by the
simple ext node spec. But instead of working on a solution to the specific
problem we tried to prop up a reasonable abstraction to cover our special
needs, and created a monster on the way.

For the simple case of adding nice syntax for your custom go model, the
ext node code is really great. The layla package stands a testament to
that, but it is just a simple layout tree.

The dom specs however must integrate with module system and resolves
references to other parts of the same declaration. These requirements are
better met with a custom implementation that reuses useful parts of the
ext node spec.